+++ ca/django_ca/tests/tests_utils.py
@@ -770,21 +770,65 @@ class ParseHashAlgorithm(TestCase):
 class FormatNameTestCase(TestCase):
     """Test :py:func:`django_ca.utils.format_name`."""
 
+    def assertFormatParse(self, value: str) -> None:  # pylint: disable=invalid-name
+        """Test formatting and then parsing again the given value as common name."""
+        name = x509.Name([x509.NameAttribute(NameOID.COMMON_NAME, value)])
+        self.assertEqual(name, x509_name(format_name(name)))
+
+        # Same, but with a different value in front
+        name = x509.Name(
+            [
+                x509.NameAttribute(NameOID.COUNTRY_NAME, "AT"),
+                x509.NameAttribute(NameOID.COMMON_NAME, value),
+            ]
+        )
+        self.assertEqual(name, x509_name(format_name(name)))
+
+        # Same, but with a different value at the end
+        name = x509.Name(
+            [
+                x509.NameAttribute(NameOID.COMMON_NAME, value),
+                x509.NameAttribute(NameOID.EMAIL_ADDRESS, "user@example.com"),
+            ]
+        )
+        self.assertEqual(name, x509_name(format_name(name)))
+
+        # Same, but with values both before and after the value in question
+        name = x509.Name(
+            [
+                x509.NameAttribute(NameOID.COUNTRY_NAME, "AT"),
+                x509.NameAttribute(NameOID.COMMON_NAME, value),
+                x509.NameAttribute(NameOID.EMAIL_ADDRESS, "user@example.com"),
+            ]
+        )
+        self.assertEqual(name, x509_name(format_name(name)))
+
     def test_x509(self) -> None:
         """Test passing a x509.Name."""
         subject = "/C=AT/ST=Vienna/L=Vienna/O=O/OU=OU/CN=example.com/emailAddress=user@example.com"
         name = x509.Name(
             [
-                x509.NameAttribute(x509.NameOID.COUNTRY_NAME, "AT"),
-                x509.NameAttribute(x509.NameOID.STATE_OR_PROVINCE_NAME, "Vienna"),
-                x509.NameAttribute(x509.NameOID.LOCALITY_NAME, "Vienna"),
-                x509.NameAttribute(x509.NameOID.ORGANIZATION_NAME, "O"),
-                x509.NameAttribute(x509.NameOID.ORGANIZATIONAL_UNIT_NAME, "OU"),
-                x509.NameAttribute(x509.NameOID.COMMON_NAME, "example.com"),
-                x509.NameAttribute(x509.NameOID.EMAIL_ADDRESS, "user@example.com"),
+                x509.NameAttribute(NameOID.COUNTRY_NAME, "AT"),
+                x509.NameAttribute(NameOID.STATE_OR_PROVINCE_NAME, "Vienna"),
+                x509.NameAttribute(NameOID.LOCALITY_NAME, "Vienna"),
+                x509.NameAttribute(NameOID.ORGANIZATION_NAME, "O"),
+                x509.NameAttribute(NameOID.ORGANIZATIONAL_UNIT_NAME, "OU"),
+                x509.NameAttribute(NameOID.COMMON_NAME, "example.com"),
+                x509.NameAttribute(NameOID.EMAIL_ADDRESS, "user@example.com"),
             ]
         )
         self.assertEqual(format_name(name), subject)
+
+    def test_escaping(self) -> None:
+        """Test various edge cases when quoting/unquoting strings."""
+        self.assertFormatParse("with/slash")
+        self.assertFormatParse('with"double-quote')
+        self.assertFormatParse("with'single-quote")
+        self.assertFormatParse("both'single\"double-quotes")
+        self.assertFormatParse("everything: slash/quote'double\"and\\backslash")
+        self.assertFormatParse('no single-quote: slash/double"quote')
+        self.assertFormatParse('no single-quote but with backslash: slash/double"quote\\backslash')
+        self.assertFormatParse("multiple\\\\backslash")
 
     def test_deprecated(self) -> None:
         """Test passing a deprecated list."""
@@ -1029,13 +1073,13 @@ class X509NameTestCase(TestCase):
 
     name = x509.Name(
         [
-            x509.NameAttribute(x509.NameOID.COUNTRY_NAME, "AT"),
-            x509.NameAttribute(x509.NameOID.STATE_OR_PROVINCE_NAME, "Vienna"),
-            x509.NameAttribute(x509.NameOID.LOCALITY_NAME, "Vienna"),
-            x509.NameAttribute(x509.NameOID.ORGANIZATION_NAME, "O"),
-            x509.NameAttribute(x509.NameOID.ORGANIZATIONAL_UNIT_NAME, "OU"),
-            x509.NameAttribute(x509.NameOID.COMMON_NAME, "example.com"),
-            x509.NameAttribute(x509.NameOID.EMAIL_ADDRESS, "user@example.com"),
+            x509.NameAttribute(NameOID.COUNTRY_NAME, "AT"),
+            x509.NameAttribute(NameOID.STATE_OR_PROVINCE_NAME, "Vienna"),
+            x509.NameAttribute(NameOID.LOCALITY_NAME, "Vienna"),
+            x509.NameAttribute(NameOID.ORGANIZATION_NAME, "O"),
+            x509.NameAttribute(NameOID.ORGANIZATIONAL_UNIT_NAME, "OU"),
+            x509.NameAttribute(NameOID.COMMON_NAME, "example.com"),
+            x509.NameAttribute(NameOID.EMAIL_ADDRESS, "user@example.com"),
         ]
     )